Separate EnhancedResolveCache into lib/enhanced-resolve-cache.js
